add button styles
