+++ src/library/Buttons/ButtonMonoInvert/index.tsx
@@ -3,29 +3,32 @@ SPDX-License-Identifier: GPL-3.0-only */
 
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { motion } from 'framer-motion';
-import '@polkadot-cloud/core/css/buttons/ButtonTertiary/index.css';
+import '@polkadot-cloud/core/css/buttons/ButtonMonoInvert/index.css';
 import type { ComponentBaseWithClassName } from 'types';
-import type { ButtonCommonProps, ButtonIconProps } from './types';
-import { onMouseHandlers } from './Utils';
-import { appendOrEmpty } from '@polkadot-cloud/utils';
+import type { ButtonCommonProps, ButtonIconProps } from '../types';
+import { onMouseHandlers } from '../Utils';
+import { appendOr, appendOrEmpty } from '@polkadot-cloud/utils';
 
-export type ButtonTertiaryProps = ComponentBaseWithClassName &
+export type ButtonMonoProps = ComponentBaseWithClassName &
   ButtonIconProps &
   ButtonCommonProps & {
+    // large button, small otherwise.
+    lg?: boolean;
     // button text.
     text: string;
   };
 
 /**
- * @name ButtonTertiary
- * @description Tertiary button style used within the main interface of dashboards.
+ * @name ButtonMonoInvert
+ * @description Inverted monotone button style used within the main interface of dashboards.
  */
-export const ButtonTertiary = ({
+export const ButtonMonoInvert = ({
   disabled,
   grow,
   iconLeft,
   iconRight,
   iconTransform,
+  lg,
   marginLeft,
   marginRight,
   marginX,
@@ -36,16 +39,17 @@ export const ButtonTertiary = ({
   onMouseOver,
   onMouseMove,
   onMouseOut,
-}: ButtonTertiaryProps) => (
+}: ButtonMonoProps) => (
   <motion.button
     whileHover={{ scale: !disabled ? 1.02 : 1 }}
     whileTap={{ scale: !disabled ? 0.98 : 1 }}
-    className={`btn-tertiary${appendOrEmpty(grow, 'grow')}${appendOrEmpty(
-      marginRight,
-      'm-right'
-    )}${appendOrEmpty(marginLeft, 'm-left')}${appendOrEmpty(marginX, 'm-x')}${
-      className ? ` ${className}` : ''
-    }`}
+    className={`btn-mono-invert${appendOr(lg, 'lg', 'sm')}${appendOrEmpty(
+      grow,
+      'grow'
+    )}${appendOrEmpty(marginRight, 'm-right')}${appendOrEmpty(
+      marginLeft,
+      'm-left'
+    )}${appendOrEmpty(marginX, 'm-x')}${className ? ` ${className}` : ''}`}
     style={style}
     type="button"
     disabled={disabled}
