+++ src/library/Buttons/ButtonOption/index.tsx
@@ -5,9 +5,9 @@ import { faChevronRight } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import '@polkadot-cloud/core/css/buttons/ButtonOption/index.css';
 import type { ComponentBaseWithClassName } from 'types';
-import type { ButtonCommonProps } from './types';
+import type { ButtonCommonProps } from '../types';
 import { appendOrEmpty } from '@polkadot-cloud/utils';
-import { onMouseHandlers } from './Utils';
+import { onMouseHandlers } from '../Utils';
 
 export type ButtonOptionProps = ComponentBaseWithClassName &
   ButtonCommonProps & {
