+++ src/library/Buttons/ButtonPrimary/index.tsx
@@ -3,15 +3,17 @@ SPDX-License-Identifier: GPL-3.0-only */
 
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { motion } from 'framer-motion';
-import '@polkadot-cloud/core/css/buttons/ButtonMono/index.css';
+import '@polkadot-cloud/core/css/buttons/ButtonPrimary/index.css';
 import type { ComponentBaseWithClassName } from 'types';
-import type { ButtonCommonProps, ButtonIconProps } from './types';
 import { appendOr, appendOrEmpty } from '@polkadot-cloud/utils';
-import { onMouseHandlers } from './Utils';
+import type { ButtonCommonProps, ButtonIconProps } from '../types';
+import { onMouseHandlers } from '../Utils';
 
-export type ButtonMonoProps = ComponentBaseWithClassName &
+export type ButtonPrimaryProps = ComponentBaseWithClassName &
   ButtonIconProps &
   ButtonCommonProps & {
+    // use secondary network color.
+    colorSecondary?: boolean;
     // large button, small otherwise.
     lg?: boolean;
     // button text.
@@ -19,10 +21,11 @@ export type ButtonMonoProps = ComponentBaseWithClassName &
   };
 
 /**
- * @name ButtonMono
- * @description Monotone button style used within the main interface of dashboards.
+ * @name ButtonPrimary
+ * @description Primary button style used within the main interface of dashboards.
  */
-export const ButtonMono = ({
+export const ButtonPrimary = ({
+  colorSecondary,
   disabled,
   grow,
   iconLeft,
@@ -39,14 +42,14 @@ export const ButtonMono = ({
   onMouseOver,
   onMouseMove,
   onMouseOut,
-}: ButtonMonoProps) => (
+}: ButtonPrimaryProps) => (
   <motion.button
     whileHover={{ scale: !disabled ? 1.02 : 1 }}
     whileTap={{ scale: !disabled ? 0.98 : 1 }}
-    className={`btn-mono${appendOr(lg, 'lg', 'sm')}${appendOrEmpty(
-      grow,
-      'grow'
-    )}${appendOrEmpty(marginRight, 'm-right')}${appendOrEmpty(
+    className={`btn-primary${appendOr(lg, 'lg', 'sm')}${appendOrEmpty(
+      colorSecondary,
+      'secondary-color'
+    )}${appendOrEmpty(grow, 'grow')}${appendOrEmpty(marginRight, 'm-right')}${appendOrEmpty(
       marginLeft,
       'm-left'
     )}${appendOrEmpty(marginX, 'm-x')}${className ? ` ${className}` : ''}`}
