+++ src/library/Buttons/ButtonTab/index.tsx
@@ -3,9 +3,9 @@
 
 import '@polkadot-cloud/core/css/buttons/ButtonTab/index.css';
 import type { ComponentBaseWithClassName } from 'types';
-import type { ButtonCommonProps } from './types';
+import type { ButtonCommonProps } from '../types';
 import { appendOrEmpty } from '@polkadot-cloud/utils';
-import { onMouseHandlers } from './Utils';
+import { onMouseHandlers } from '../Utils';
 
 export type ButtonTabProps = ComponentBaseWithClassName &
   ButtonCommonProps & {
