+++ src/theme/fonts.scss
@@ -5,26 +5,26 @@ SPDX-License-Identifier: GPL-3.0-only */
   font-family: Inter;
   font-style: normal;
   font-weight: 600;
-  src: url("./fonts/Inter600.woff2") format("woff2");
+  src: url('./fonts/Inter600.woff2') format('woff2');
 }
 
 @font-face {
   font-family: InterSemiBold;
   font-style: normal;
   font-weight: 700;
-  src: url("./fonts/Inter700.woff2") format("woff2");
+  src: url('./fonts/Inter700.woff2') format('woff2');
 }
 
 @font-face {
   font-family: InterBold;
   font-style: normal;
   font-weight: 800;
-  src: url("./fonts/Inter800.woff2") format("woff2");
+  src: url('./fonts/Inter800.woff2') format('woff2');
 }
 
 @font-face {
   font-family: Unbounded;
   font-style: normal;
   font-weight: 800;
-  src: url("./fonts/Unbounded.woff2") format("woff2");
+  src: url('./fonts/Unbounded.woff2') format('woff2');
 }
